Add files via upload
This repository contains a static resume built using HTML, CSS, and TypeScript for Hakathon challenge. The project demonstrates expertise in front-end development and presents a professional online resume. It aims to provide a responsive and visually appealing platform for showcasing personal and professional details.
Hackathon Challenge:
Our project tackles the static resume is presented in hackathon challenge ,  It highlights an innovative approach to personal branding through a static resume designed to be both functional and aesthetically pleasing.